pyproject.toml
[project]
name = "kth-dataset-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.19.11",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pathlib>=1.0.1",
    "pypcd4>=1.4.3",
    "pyzmq>=27.1.0",
    "ruamel-yaml>=0.19.1",
    "seaborn>=0.13.2",
]
